split persist node values and update dependencies logic
