--- src/server/controller/cycleData/persistNodeValues/calculateDependantNodes/index.ts
@@ -1,1 +1,0 @@
-export { calculateDependantNodes } from './calculateDependantNodes'
