+++ src/server/controller/cycleData/persistNodeValues/index.ts
@@ -1,1 +1,2 @@
 export { persistNodeValues } from './persistNodeValues'
+export type { PersistNodeValuesProps } from './props'
