--- src/server/controller/cycleData/persistNodeValues/validateNodeUpdates/index.ts
@@ -1,1 +1,0 @@
-export { validateNodeUpdates } from './validateNodeUpdates'
