+++ src/server/controller/cycleData/updateDependencies/expressionEvaluator/functions/validatorEqualToTotalGrowingStock.ts
