+++ src/server/controller/cycleData/updateDependencies/props.ts
@@ -1,0 +1,3 @@
+import { PersistNodeValuesProps } from '@server/controller/cycleData/persistNodeValues'
+
+export type UpdateDependenciesProps = PersistNodeValuesProps & { isODP?: boolean }
