+++ src/server/controller/cycleData/updateDependencies/updateValidationDependencies/index.ts
@@ -1,0 +1,1 @@
+export { updateValidationDependencies } from './updateValidationDependencies'
