+++ src/server/controller/cycleData/updateDependencies/updateValidationDependencies/validateNode.ts
@@ -2,8 +2,8 @@ import { NodeValueValidation, NodeValueValidations, Row } from '@meta/assessment
 import { TableData } from '@meta/data'
 
 import { getTableData } from '@server/controller/cycleData/getTableData'
-import { ExpressionEvaluator } from '@server/controller/cycleData/persistNodeValues/expressionEvaluator'
 import { PersistNodeValueProps } from '@server/controller/cycleData/persistNodeValues/props'
+import { ExpressionEvaluator } from '@server/controller/cycleData/updateDependencies/expressionEvaluator'
 import { BaseProtocol } from '@server/db'
 
 export const validateNode = async (
